Add explicit type conversions
This is in preparation for turning on -Wconversion:
- fastpath.h: Cast to register_t from a + b is safe
- fastpath.h: Cast to hw_asid_t is safe from bitfield.
- machine.h: Change clearMemory to take word_t type.
  This is aligned with arm.
- util.c: int cast to match up with compiler builtins
- util.c: long cast in str_to_long()
- fastpath.c: syscall_t cast for negative syscall numbers
- tcb.c: intentionally cast to breakpoint number type.
- untyped.c: Retype chunk to word_t.
- untyped.c: Cast loop iterator to sword_t and then word_t
  to handle the loop condition breaking when decremented
  past 0.

+++ src/object/tcb.c
@@ -558,7 +558,7 @@ static exception_t decodeConfigureSingleStepping(cap_t cap, bool_t call, word_t 
 
     tcb = TCB_PTR(cap_thread_cap_get_capTCBPtr(cap));
 
-    bp_num = getSyscallArg(0, buffer);
+    bp_num = (uint16_t)getSyscallArg(0, buffer);
     n_instrs = getSyscallArg(1, buffer);
 
     syserr = Arch_decodeConfigureSingleStepping(tcb, bp_num, n_instrs, false);
@@ -588,7 +588,7 @@ static exception_t decodeSetBreakpoint(cap_t cap, word_t *buffer)
     syscall_error_t error;
 
     tcb = TCB_PTR(cap_thread_cap_get_capTCBPtr(cap));
-    bp_num = getSyscallArg(0, buffer);
+    bp_num = (uint16_t)getSyscallArg(0, buffer);
     vaddr = getSyscallArg(1, buffer);
     type = getSyscallArg(2, buffer);
     size = getSyscallArg(3, buffer);
@@ -724,7 +724,7 @@ static exception_t decodeGetBreakpoint(cap_t cap, bool_t call, word_t *buffer)
     syscall_error_t error;
 
     tcb = TCB_PTR(cap_thread_cap_get_capTCBPtr(cap));
-    bp_num = getSyscallArg(0, buffer);
+    bp_num = (uint16_t)getSyscallArg(0, buffer);
 
     error = Arch_decodeGetBreakpoint(tcb, bp_num);
     if (error.type != seL4_NoError) {
@@ -751,7 +751,7 @@ static exception_t decodeUnsetBreakpoint(cap_t cap, word_t *buffer)
     syscall_error_t error;
 
     tcb = TCB_PTR(cap_thread_cap_get_capTCBPtr(cap));
-    bp_num = getSyscallArg(0, buffer);
+    bp_num = (uint16_t)getSyscallArg(0, buffer);
 
     error = Arch_decodeUnsetBreakpoint(tcb, bp_num);
     if (error.type != seL4_NoError) {
